refactor: reorganize SalaryWorkTime file structure
- Rename TimeAndSalary → SalaryWorkTime to reflect actual domain
- Move OvertimeBlock into OvertimeSection/ (only used there)
- Move formatter, injectHideContentBlock, InfoModal into WorkingHourTable/ (only used there)
- Move WorkingHourBlock.module.css next to WorkingHourBlock.js
- Remove stale src/components/SalaryWorkTime/common/ directory
