+++ src/components/CompanyAndJobTitle/SalaryWorkTime/EsgBlock/EsgBlock.js
