+++ src/components/CompanyAndJobTitle/SalaryWorkTime/EsgBlock/EsgBlockDesktop.js
