+++ src/components/CompanyAndJobTitle/SalaryWorkTime/EsgBlock/EsgBlockMobile.js
