+++ src/components/CompanyAndJobTitle/SalaryWorkTime/OvertimeSection/index.js
@@ -1,0 +1,3 @@
+import OvertimeSection from './OvertimeSection';
+
+export default OvertimeSection;
