+++ src/components/CompanyAndJobTitle/SalaryWorkTime/SummarySection.js
