+++ src/components/CompanyAndJobTitle/SalaryWorkTime/WorkingHourTable/InfoModal.module.css
