+++ src/components/CompanyAndJobTitle/SalaryWorkTime/WorkingHourTable/index.js
@@ -1,0 +1,3 @@
+import WorkingHourTable from './WorkingHourTable';
+
+export default WorkingHourTable;
