+++ src/pages/Company/CompanySalaryWorkTimeProvider.js
@@ -10,8 +10,7 @@ import {
   queryRatingStatistics,
 } from 'actions/company';
 import { paramsSelector, querySelector } from 'common/routing/selectors';
-import { useSearchTextFromQuery } from 'components/CompanyAndJobTitle/SearchBar';
-import SalaryWorkTime from 'components/CompanyAndJobTitle/TimeAndSalary';
+import SalaryWorkTime from 'components/CompanyAndJobTitle/SalaryWorkTime';
 import {
   dataTimeFromQuerySelector,
   experienceFromQuerySelector,
@@ -23,7 +22,8 @@ import {
   useExperienceFromQuery,
   useGenderFromQuery,
   useSortByFromQuery,
-} from 'components/CompanyAndJobTitle/TimeAndSalary/SalaryFilter';
+} from 'components/CompanyAndJobTitle/SalaryWorkTime/SalaryFilter';
+import { useSearchTextFromQuery } from 'components/CompanyAndJobTitle/SearchBar';
 import { PAGE_SIZE, PageType, TabType } from 'constants/companyJobTitle';
 import { usePage } from 'hooks/routing/page';
 import usePermission from 'hooks/usePermission';
